Fixed Jar Resolver auto-resolution.
There were some cases where auto-resolution wouldn't happen correctly
now the Version Handler is part of the plug-in and plays with asset
metadata on load.  So this change simplifies auto-resolution to only
prevent re-entrant auto-resolution and also initializes the app bundle
ID on load so that an auto-resolution doesn't occur the first time
the bundle ID is polled.

Also:
* Fixed stale bundle ID used to perform resolution when it's changed.
* Restored version handling of the plugin's manifest.

BUG=30116788

Change-Id: I56e035a023625e63e20d422778cbb18ac3ddc9e0