Fixed VersionHandler and IOSResolver compatibility with Unity 4.
Unity 4 doesn't have the UnityEngine.PluginImporter class and
stores the Xcode DLL in a yet another location.

Since the PluginImporter is not available in Unity 4,
VersionHandler targetting is disabled as it's not possible
- without hacking .meta files - to enable / disable targeting
for plugins (DLLs).

As the Mono distribution is slightly different, the IOSResolver
has been modified to only call methods that use the Xcode
plugin in separate methods after the workaround for loading
the plugin has been installed.  In addition, since the Xcode DLL
seemingly moves with each Unity OSX release, a brute force file
search is now used to search for the Xcode DLL if it isn't in
the default location.

BUG=31336359

Change-Id: Ia45055abd460b10a018f3a1623833a558283e646