Updating build script to use Unity distro only.
This is done to facilitate CI builds.

This removes the Dockerfile and monodevelop build environment.

Introduced properties in the csproj files that are used to define the
directory for the Unity and NUnit assemblies.  These are set in the build.gradle
file by searching the paths in Unity.

If you are building the solution outside the gradle script, you'll need to
update the hints manually in the IDE.

Change-Id: I1282dcf50c0efdee2e69df04ab207bf1f1f67966