Add views for package management
This change introduces the views used to manage project plugins,
settings and registries. A new controller is introduced along with
the views that performs the plugin installs. Plugin unistall has
been stubbed out. Some changes were made to controllers in order
to support needed features for the views to work.

Change-Id: Ibc956d49f381f2ee07e6fb330cbb348dc561bab2