Fixed IOSResolver to handle PlayerSettings.iOS.targetOSVersionString
PlayerSettings.iOS.targetOSVersionString now specifies the target iOS
SDK.  This is problematic as the new string is freeform and can
therefore result in users providing an invalid version which can cause
builds to fail.  In the case of the IOSResolver plugin it was causing
an exception when trying to convert the string to an integer
representation.  This patch applies the following:

* Uses reflection to choose between targetOSVersionString and
  targetOSVersion (enum) methods of reading / writing the target SDK
  verion.
* Provides validation of the version string when converting to an
  integer representation.
* Handles invalid conversions (though that shouldn't happen if the
  validation is correct).
* Provides a fallback / default version when the provided SDK string
  can't be parsed.

BUG=33635555

Change-Id: Ief689cd887ab75b7d3c9d78ea9f3c349d8a7d59c