Handle exceptions thrown by Unity.Cecil on asset rename
When the VersionHandler activates a package it potentially renames
DLLs to the canonical name causing the DLL to be reimported.  In
some cases this can cause Unity.Cecil (as far I can tell they use this
to rewrite DLLs) to throw exceptions while processing the DLL which
are totally harmless as they DLL is imported and processed again
when AssetDatabase.Refresh() is executed.

BUG=34123612

Change-Id: I20e993aa0de8ff8b1075258d042ccf076b945ddf