Refactor Package Manager to use xml dependency declaration
Changed the packaging model for packaged plugins so that they now use
xml to declare all dependencies. This change allows for labeling of the
resolved dependencies and clean plugin removals even if the transitive
dependency graph changes due to external library versioning.

Install and Uninstall of packaged plugins has changed.

Change-Id: I8ec5394e01f3d4c0104bac11e027cc6cc7ae66a0