Handle empty string returned by Unity for "targetOSVersionString" property.
New projects in Unity can sometimes return an empty string for the
iOS targetOSVersionString property. Toggling the value within Unity
build settings will resolve this issue. If not toggled, podfiles will
be created with '' as the target sdk and pod install commands will fail
as the minimum target criteria of pods will not be met. Checking for both
and empty string will default to using DEFAULT_TARGET_SDK in this
scenario.

Change-Id: I4c618e9b04c9e671247ffaa42be6017200e4e670