Added a script that uses Gradle to build a set of dependencies.
This is needed in order to add proguard support within Unity. Unity doesn't
expose their internal android build pipeline to hook in proguard or offer
support for it, and yet on android when linking to so many libraries there's a
major issue with hitting dex limits.

The script handles generating a complete gradle project, with a configurable set
of dependencies. We then build the project using proguard and extract the useful
parts to a folder in an Eclipse / Ant format (which can be used with Unity).

This approach allows us to re-use many of the prerequisite steps and
functionality built into Gradle and the Gradle Android plugin. By triggering the
proguard step, we automatically get resource compilation, dependency resolution,
macro expansion, proguard config aggregation, and other prerequisite,
without replicating the toolchain.

Change-Id: Ifbe55b0963c7694917db015c13e1d8991c87a6a9