Added option to run pod tool via the shell.
Some installations require customizations in .bash_profile or
.bashrc to correctly execute the Cocoapods command line tool.  This
adds an option to the iOS resolver settings to run pod tool via the shell
environment rather than simply forking process to run the executable.

Shell execution is disabled by default as it's not possible to redirect
stdout / stderr streams with the mode enabled which means that it's not
possible to report any details of errors if they occur.

Bug: 32786297
Change-Id: Idaf79a395612c22b05f69ef8434d81446d4a1fc4