Added support for individual ABIs when using AARs with native libs.
Unity's legacy system includes all ABIs from AARs when targeting
a single ABI.  This results in APK size bloat and also can break
plugins when targeting x86 platforms that support dynamic
recompilation.

This commit adds an Android target ABI poller which triggers
an event when the user changes the ABI in the editor.  If the
ABI changes the resolver is notified and potentially re-processes
the AARs included in the project to strip the native libraries that
are not required by the target ABI.

The behavior varies by Unity version:
* Unity 4:
  - Always unpacks AARs, always targets the FAT ABI.
* Unity 5 (Legacy Build):
  - Unpacks AARs if they contain native libraries and a single ABI
    is selected.
  - Removes ABIs that are not required for the selected build target.
* Unity 5 (Gradle Build):
  - Same as Unity 5 legacy build.  Unity's Gradle integration - similar
    to the legacy build blindly includes unneccesary ABIs.

In addition, this change fixes synchronization of the AAR explode cache.
AARs are now indexed by package name and removed from the cache when
they're stale.

Bug: 33073014
Change-Id: Id94d3e0b93a826e597294833037594cba04ebab3