Workaround Cocoapods installation when using Ruby < 2.2.2
Cocoapods requires the activesupport gem (Ruby package).  However,
the most recent release of the activesupport gem requires
Ruby >= 2.2.2.  If activesupport isn't installed, this patch
installs activesupport 4.2.6 which is compatible with earlier
version of Ruby.

In addition, this patch changes Gem installation to only install
in the user's directory using --user-install if RVM is not installed.
RVM configures the Gem package manager to install in a separate
user subdirectory for each revision of Ruby installed on a machine.

Bug: 33181661, 31709520
Change-Id: If1ee9c6b98efad6355cb0a830075e8dad918b3a2