Workaround for package resolution when using Google Play Services.
The Jar Resolver doesn't evaluate the entire dependency graph when
evaluating dependencies which leads it to choose the latest version of
a package without re-evaluating all constraints associated with the
package.  This poses a problem for Google Play Services dependencies
which all need to be pinned to the same version to function correctly.

Rather than implementing a replacement dependency resolution system -
which could take a while - this patch adds the concept of "version
locking" to pin all packages within a group to a particular version.
This is applied to Google Play Services packages such that a
consistent, compatible version is selected across the package set.

In addition, this patch adds support for verbose logging in the
the PlayServicesSupport component of the Jar Resolver with an
option to enable / disable the functionality.

Bug: 33354806

Change-Id: Ie10657143742685eca2832f177a2efc7fd518286