Persisted choice to ignore AAR / JAR version replacement.
If a conflicting Android dependency (AAR / JAR) is resolved and
package already exists in the target Plugins/Android directory the
user is asked whether to replace the package or leave it.  Previously
the choice to leave the package would result in the user being
prompted each time resolution occurs.  With this patch, the ignored
package version is stored in the AAR explode cache so that the user
isn't asked about replacement until the next time they open the
editor.

This also fixes a bug which resulted in the cache not being flushed
to disk when items change in the cache but the size doesn't change.

Bug: 33948053
Change-Id: I9512df06aadb444271682aefb11ed2d80206149f