Added support to pass proguard configs to fix messaging.
The Messaging AAR has a soft dependency on Analytics which was
failing in proguard. This CL adds support for passing proguard
configuration files to the GenGradle script, and currently provides
the unity proguard config and messaging workaround proguard configs.

This also removes the unity dependency from the gradle-template.zip,
making the generate_gradle_prebuild.py unity and firebase agnostic.

Change-Id: I70109ccaddb80d16706bb6dbd3d46c8442851bf6