Fixed propagation of extra repository paths.
When a client adds a dependency and specifies additional repositories
to search, it only worked for that artifact, while transitive
dependecies of that ended up only searching the default set of
repositories. This change makes it so that extra repositories for an
artifact are also considered when searching for dependencies of that
artifact.

Change-Id: I3acc5930abc702838c01dbaf17ff0ca9aa07ee82