Restored AAR state cache from AAR files in the project.
The AAR state cache was moved into the Temp directory in change
I5a24cc1916c3cd941f5b1e80c4bf04e58211d1c6.  This results in the cache
being deleted when Unity closes which is problematic as the resolver
uses the cache to avoid touching the local maven repo when
auto-resolution is enabled.  This patch recreates the AAR state cache
when the JarResolver plugin is loaded as no cache file exists.

Bug: 37518315
Change-Id: I88a2a3646c64df7d0ff17bf807dc75065a722c2c