Fixed NullReferenceException when a dependency isn't found.
PlayServicesSupport.FindCandiate() can return a null Dependency if it isn't
present in the set of local maven repositories.  This was causing a null
reference exception in PlayServicesSupport.ResolveDependencies() when
attempting to copy the returned Dependency object.  This patch fixes the
issue by skipping missing dependencies.  An error is already reported by
FindCandidate().

Change-Id: I3bb1c0ee3ed4afc9483f9336fda3d03f16acfa1b