Added notification on VersionHandler update completion
When the VersionHandler exceutes it sometimes needs to refresh the
asset database in order to compile newly enabled DLLs.  Since asset
database refresh operations are asynchronous it's not easy for an editor
script to determine when compilation is complete.   This commit adds
functionality to the VersionHandler which allows a developer to provide
a list of static methods that are called when an update operation is
complete.  This makes it easier for build scripts to start the build
process for a target platform when the VersionHandler's project
preprocessing operations are complete.

Bug: 36842933
Change-Id: I4811ae349de5e78edbfe5dcd446c5006add3274e