Added support for Gradle based resolution when prebuild is disabled.
This patch deprecates most of PlayServicesSupport.cs now relegating the
module to speeding up the process of determining whether to execute a
resolution step.  In addition, since Gradle is now used to fetch Android
dependencies, non-local artifacts are now supported.

The legacy behavior is still in place and installation of packages via
the Android package manager is retained when using Gradle resolution.

Bug: 64454251
Change-Id: Ifceb5f28b80fde32586e17dd32c76c02821a18a2