Improved Android resolver auto-resolution and Gradle dependency cleanup
When using Gradle to resolve, stale transitive dependencies were not
being cleaned up.  In addition, auto-resolution was not efficient when
using Gradle to build.  This patch introduces a ProjectSettings file
that is used to track the set of dependencies in a project.  When the
set of dependencies declared by *Deps.cs files or the dependencies under
Plugins/Android don't match what is expected, auto resolution is triggered.

Also, fixed the Gradle Prebuild to call the completion delegate.

Resolves issue https://github.com/googlesamples/unity-jar-resolver/issues/62

Change-Id: I9357c8a6f7dcfee5d52b7a9dff6d13a50800b722