Fixed Android Gradle resolution when a space is in the path.
Change I85547aabba23fda4ebe2f2cb81e589f5b8e635dc modified the path
to the download_artifacts.gradle script such that it's now an absolute path.
This means that path to the Unity project is added to the gradle command line.
When the path contains whitespace the script is no longer correctly passed to
gradle.  This commit quotes the path to fix this issue.

Change-Id: Ia6fd0bf910240b5c00957acc36daac4c5cd6a2ea