Fixed recursive polling job execution.
It's possible for a polling job to schedule additional polling tasks.  In batch
mode, since everything runs on the main thread with multiple invocations of
ExecutePollingJobs(), instances of ExecutePollingJobs() up the stack can be in a
state where the job iterator ends up pointing outside of the job list leading to
an OutOfRangeException.

This commit changes ExecutePollingJobs() to stop executing jobs when the end of
the job list is reached.  In addition, this changes PollOnUpdateUntilComplete()
to only poll until the currently scheduled job is complete in batch mode to
prevent waiting for all jobs to finish.

Bug: 79148523, 77549386
Change-Id: I3e282231f6f6b96aa2f4843c37941502021f402b