Attempted fix for race when reading / writing explode cache.
The Android Resolver uses a temporary file to cache information about AARs /
JARs that can be preserved across recompilation which forces DLL reloads.
Methods used to read / write the cache were not in using statements which could
result in file handles being left open after they're read written.  In addition,
if a file operation failed for this optimization, the Android Resolver would
fail to load or stop working.

This wraps file operations in using statements so file handles are closed when
the read / write completes and wraps the file read / write operations in
exception handling so that the resolver doesn't completely fail when accessing
the cache.

Fixes #141
Bug: 112652392

Change-Id: I5c180da49a8e17d9fff62d6e06d236b835d2b170