Fixed Android Resolution when the build target isn't Android.
If the Android build target is selected in the Unity editor then
the user selects a different target using the build selection dialog
and then runs the Android resolution process, the wrong applicationId
is retrieved in Unity 5.6 and above and patched into AARs.

This fixes the issue by explicitly retriving the applicationId of the
Android build target rather than the target the user has selected in
the editor.

Bug: 62466035
Change-Id: Ic6dd524b8d41846381481c17c8cb4dfad4b4a83e