Fix ScheduleAutoResolve() triggering AutoResolve() twice.
ScheduleAutoResolve() will schedule an AutoResolve() via a the lambda
for the function RunOnMainThread.PollOnUpdateUntilComplete(). The
lambda can potentially be called again after AutoResolve() has
completed the resolve, resulting in another auto-resolution occuring.

Fixed by checking the autoResolveJobId that is created when scheduling
the AutoResolve(). If autoResolveJobId is 0 then we can assume the
resolution has been scheduled and we do not call AutoResolve() again.

Bug: 118340254
Change-Id: Ia62a298f71aa96e835602669859f33a767cc1a3c