Only perform auto-resolution on the first scene while building.
The Android Resolver was running the resolution process for every
scene included in a build which could significantly extend build times.
Even worse, OnPostProcessScene is called when starting the player in the
editor which would slow down player startup times.

This changes the synchronous resolution process to only run if a build
is in progress and also ensures resolution only run for the first scene
in the project.

Fixes #170
Fixes #167
Bug: 121155647

Change-Id: I38e4f2653c339e0c4fd2dfc62cd53ca20fceb4ef