Reduced Android Resolver property polling frequency.
Even though PropertyPoller had a checkIntervalInSeconds property, the
Poll() method required the current value of the property which ended up
polling the propety *every* frame rather than every checkIntervalInSeconds.
This changes the Poll() method to accept a delegate that retrieves the
current value instead of requiring the value on every call to the method.

Fixes #146
Bug: 121162585

Change-Id: I7e5f8ec2967ad9f1b530173da4c2e108b6e6b890