Changed some file operations to workaround Windows permission issues.
The Android Resolver processes AARs and JARs in temporary directories
moving the results to the Unity project when it's complete.

Some users (e.g firebase/quickstart-unity#295 ) have reported that
moving files or directories from the temporary folder can result in
exceptions due to lack of permissions.

I've been unable to reproduce the user's issue. However, some sources
on the web suggest that old .NET versions of File.Move() and
Directory.Move() do not correctly propagage permissions on Windows.

To workaround broken implementations of File.Move() and Directory.Move()
on Windows this patch does the following:
* Moves all temporary files under the Unity project's Temp directory
  instead of the system temporary folder.  This folder should be writable
  by the user otherwise basic operations in Unity would fail.
* Changes File.Move() and Directory.Move() operations to copy and delete
  on Windows so that permissions of files and folders in the target location
  inherit from the Unity project folder.

In addition, this modifies ProcessAar() to log an error but continue when an
error occurs so that a modal progress bar isn't left displayed for the Unity
session.

Change-Id: Ib55ef90d2b8418fb262ea50756600f438a0b91a6
GitHub: firebase/quickstart-unity/issues/295
Bug: 123421567