Add Enable/Disable prompt before the first Android auto-resolution
Create a modal dialog that will display when the Android Resolver is about to run for the
first time to let the user choose whether to enable or disable auto-resolution at the
outset. If disabled, warn the user via console log at build time to ensure that they
have run the Android Resolver, and show them where to manually run or re-enable auto-resolve.

Change-Id: I2bfdc91a16597cdb5711418cd4083b723d0ece9b