Only warn if the console encoding cannot be changed.
The console encoding needs to be set to UTF8 to run Cocoapods
command line tools and pipe data to the Android SDK manager.  In
some environments it's not possible to change the encoding from the
default so warn the user instead of throwing an exception and failing
to run the command entirely since the command will report an error to
Unity's log output if it fails.

Fixes #300
Bug: 127302270

Change-Id: I44773ca539b7723f6c4e6bb8b18add4e65108032