Removed use of EditorUtility.DisplayProgressBar in Android Resolver
EditorUtility.DisplayProgressBar() is very flakey, it can end up in odd
states where it can never be cleared.

This removes the use of EditorUtility.DisplayProgressBar from the Android
Resolver, instead moving the progress bar display to the CommandLineDialog.
In addition, long running operations like labeling assets, have been made
asynchronous by scheduling each iteration on the main thread.

Finally, this leaves the Android Resolver window open if resolution fails,
results in one or more warnings or verbose logging is enabled so that it's
easier to inspect the failure logs.

Bug: 132070706
Change-Id: I31ed32659d3871f501aab7305cbfe4df99613e89