Fixed Android Resolver log spam on load.
The Android Resolver was loading the AAR cache in a static constructor
and using the log functions before they were initialized with the
log level from project settings.  It was possible to configure system
wide logging as "verbose" - which would be used in static constructors -
before the project settings are loaded.

Fixes #192

Bug: 132071158
Change-Id: Icdd5d5d12dfbf4d656885360bfceca15da2f7008