Support for Unity 2019 Android SDK and JDK install locations
Unity 2019 adds support for installation of the Android SDK and JDK in
the AndroidPlayer directory making it easier for developers to configure
the Android build environment.  Unfortunately, Unity does not update
the JdkPath and AndroidSdkRoot editor preferences when the Android SDK
and JDK are installed with the AndroidPlayer.

This change searches for the Android SDK and JDK if JdkPath and
AndroidSdkRoot are not set.

Fixes #188

Bug: 132072681
Change-Id: If01f642ff81570d9c28552fab6957bc88cf4dac1