Fixed up maven repos embedded in Unity projects when in Gradle builds.
When resolving Android dependencies using Gradle and a mainTemplate.gradle
file, srcaar artifacts in maven repos embedded in the Unity project are
renamed to aar so they can be included in the Android build process.

However, some plugins (like Google Play Games) ship with Maven POMs that
refrence the srcaar artifacts directly rather than the aar artifacts.
This results in the renamed aar files being ignored in the Android build
process.

This commit patches Maven POMs so they reference the correct artifact
when the packaging is changed by the Gradle Android Resolution step.
If libraries are removed, the POMs are updated to point at the packaging
for the artifact in the directory (e.g srcaar).

Fixes #208
Bug: 132168073

Change-Id: I357c6a1197d4fc15cc924b5a4d9e4d7c79be922d