Fixed Gradle template builds in Unity 2019
Unity 2019 changes the developer's Android project to a library when built by
Gradle.  The application project is now a launcher which includes this library.
However, when the application is built it doesn't include the repositories of
the developer's library project causing any repos added by the Android Resolver
to be ignored.  This commit fixes the issue by adding all Android Resolver
specified repositories to all top level Gradle projects.

In addition, this commit adds support for relative repo paths in
mainTemplate.gradle.

Bug: 133275965
Bug: 133261012
Change-Id: I863640bada62bd671bb68ebb1469490dba280311