Added compatible way to get the AndroidPlayer directory.
BuildPipeline.GetPlaybackEngineDirectory() was added in Unity 2019
to get the playback engine directory.  The Android playback engine
directory is required to determine the version of Gradle shipped
with the plugin so this does not work in Unity < 2019.

This patch adds a compatible but slower way to retrieve the
AndroidPlayer directory which also fixes the generation of Gradle
dependency inclusion statements in Unity < 2019

Fixes #213
Bug: 133243196
Change-Id: I455ec6de28900a31ea414bd1ff657404e611b0fb