Only poll build properties for auto-resolution with Android build target.
Unity 2019.x throws an exception when
BuildPipeline.GetPlaybackEngineDirectory() is called and the Android player
isn't installed.  The Android build target can't be selected if it's not
installed, so we shouldn't be polling the build system if the build target
isn't Android.  This patch disables build system property polling on assembly
load if the build target isn't Android.

Fixes #220
Bug: 134665717

Change-Id: I147e51a9c7740d2e21fdd376abd7e6b2e7cdc7dc