Fixed resolution of Android dependencies without version specifiers.
Android Resolver's artifact downloader script was failing if a
dependency in the graph did not specify a version for an artifact.
This commit makes it possible to handle empty version specifications
in dependencies.

In addition, this adds more information to conflicting package log
messages and eliminates the creation of copy tasks for output
files that already exist.

Fixes #227
Bug: 134571587

Change-Id: I81d52d40c2447b6a6bfba68621fbd4b254570850