Added Jetifier to download_artifacts.gradle
Added Jetifier support to the Gradle artifact downloader script.
download_artifacts.gradle can now substitute direct and transitive
dependencies on the legacy support libraries to Jetpack and rewrite
references to the legacy support libraries in all downloaded libraries.

This adds unit and integration tests for this functionality but does not
yet connect this to the Android Resolver Unity plugin.

https://github.com/googlesamples/unity-jar-resolver/issues/145
Bug: 113575309
Change-Id: Ibb92e43751425ebfd7cedfae62d275099860aaaa