Worked around bug in Gradle 5.1.x
Gradle now runs task execution in parallel by default but due to
https://github.com/gradle/gradle/issues/6068 it's not possible
to capture stdout / stderr from individual tasks which breaks
download_artifacts_test.gradle.  This commit modifies the test to run
all tasks serially, prepares for the Gradle bug to be fixed and
cleans up the output a little when tests fail.

https://github.com/googlesamples/unity-jar-resolver/issues/145
Bug: 113575309
Change-Id: Ifee400beaa4544b3d9145cce51b69125ffc1f5b9