Do not update asset GUIDs unless the target filename changes.
Developers tend to move the plugin around in their project. However,
the Google.VersionHandler.dll asset GUID was changing with each release
resulting in the DLL failing to load if multiple versions are present in
the project. This patch fixes this issue by only generating new GUIDs
for files that are renamed in the plugin.

Bug: 127793909
Change-Id: I4569e467f3f0c1f96722a280d62af5916344af3e