Connected Jetifier to the Android Resolver plugin.
This adds an option to enable Jetpack (AndroidX) Jetifier processing as
part of the either dependency resolution or Gradle builds.

This change also ensures that the user is notified of build environment
settings that will prevent the Jetifier from functioning or prevent the
use of the Jetpack (AndroidX) libraries.
See...
- https://developer.android.com/studio/releases/gradle-plugin#3-2-0
- https://developer.android.com/jetpack/androidx#using_androidx

https://github.com/googlesamples/unity-jar-resolver/issues/145
Bug: 135263921
Change-Id: Iddb2ab7e34b411973c761360e485aed2506f2ddc