Run Android auto-resolver when relevant settings change.
Auto-resolution was previously relying upon polling for differences in
runtime state which could break if the DLL was reloaded after settings
are changed.  This commit stores the resolution settings with the rest
of the state from the last resolution so that it's possible to
auto-resolve if any of this state changes or the settings change that
require resolution.

Bug: 135263921
Change-Id: I5858cebdb905f6425faee719032b830c9a066a81