Enforce API level 28 is selected if AndroidX is detected.
AndroidX doesn't compile with API level < 28 which results in a cryptic compile
error for developers who haven't installed API 28 in their Android SDK or
selected a lower API level for compilation in Unity. This detects AndroidX
libraries when using Gradle resolution - this isn't possible at the moment for
transitive dependencies when using the mainTemplate.gradle patcher - and
prompts the user to fix their build configuration otherwise resolution is
aborted.

This also fixes cases of multiple completions in LabelAssets and the AAR
processing code.

Bug: 136120351
Change-Id: I052c69fd8abdb8d991e5777fc64a1e3d30cf9ca9