Moved embedded resource management into a shared module.
Added the EmbeddedResource module which moves the logic to extract DLL
embedded resources out of GradleResolver so that it can be resused more
easily by other modules. This also adds functionality to keep extracted
resources up to date when the source assembly is updated and modifies
zip extraction to conditionally extract files based upon modification
time.

Bug: 135269831, 135273820
Change-Id: I259743e4ae7475a05a5140f2bc9ad6099d40be16