Remove AAR cache from GradleResolver
The GradleResolver component maintained an AAR cache which determined the
state of each resolved library and what had changed since the last
resolution.  However, as of 76e7a4434801e277d09248d469e7130409e35c61
whether to run resolution again changed to simply be a function of
whether the specified dependencies, settings or expected output
files changed.  This left the AAR cache in a state where the only value it
added was reducing the number of times a library needed to be inspected
between each resolve step.

This patch merges AAR explosion / patching with inspection to speed up
the AAR processing process which eliminates the gains of the AAR cache.
In cases of incremental resolution, the AAR would be faster but in practice
those cases are pretty rare typically only happening when a plugin in upgraded.

Bug: 138113522
Change-Id: I122aa1bc8b528554a92910342d6964c5529e6cc3