Fixed execution of methods with interface args using reflection.
GetType() returns the concrete type of an object which caused
InvokeMethod() and InvokeStaticMethod() to not match methods that
take arguments that require interface arguments.

This changes the type check of arguments, to handle overloaded
methods, to check whether an argument is assignable given an argument
which checks for compatibility with the expected interface.

Bug: 138224253
Change-Id: I575a03b10bd3b8576fb1c4503c57ddf9052587ca