Fixed iOS target SDK version check.
Fixed the iOS target SDK version check for Cocoapods to validate the
target SDK is greater than or equal to the maximum required SDK for
all Cocoapods included in a project.

This also, modifies the iOS resolver to load the pod configuration
when the plugin starts up, if dependency configuration files change
and adds validation to the iOS SDK build setting.

Bug: 134614506
Change-Id: Ia43d62a7e2c073b91040840e3b6bf08f59d60163