Fixed Unity Asset Uploader unit test.
This changes unity_asset_uploader to no longer throw an exception if
the response doesn't have a status field.  The test has been fixed to
correctly build the login response and all metadata responses.
In addition, this moves all assertions in the test to use the unittest
module so that failures are reported with useful error messages rather
than simply "assetion failed".

Change-Id: I6abb180330bb815568a7a43d4b3f7184317ab70a