Fixed reset of settings if a setting set occurs before a get.
Setting a project setting would change the in-memory settings and then
save the current state to disk which would delete all previously saved
settings.  This changes the ProjectSettings to always load settings
before mutating and saving them.

In addition, this changes VersionHandlerImpl so that running in batch
mode doesn't change the verbose logging setting.

Fixes #301
Bug: 145540238

Change-Id: I869e7a0e867fe1991ffc6ce866541adec6d2d8e3