Disable all obsolete files that are referenced by packages.
Previously, the Version Handler would only disable a file if
it was determined as obsolete on a file-by-file basis.
For example, a/b/c/foo-1.2.3.dll is older than a/b/c/foo-1.2.4.dll
so a/b/c/foo-1.2.3.dll would be disabled and marked for deletion.

However, if a file was obsolete *and* moved between package updates
it would not be disabled but would be marked for deletion as the
file would not exist in the most recent package manifest. This
scenario is problematic if the developer doesn't clean up the old
file when prompted (target builds break) or the file is loaded in
the editor (editor plugins break).

This change adds files that are determined obsolete by
comparing differences between package manifests to the set of
files to disable.

Bug: 150886091
Change-Id: I49c711feef9817c79e9d36ffebc4f8503fba33ed