Ensure FileMetadata does not attempt to update UPM files.
FileMetadata synchronizes asset labels with properties on the
class which does not work with UPM as UPM will restore the files
after modification, cause an asset database refresh which then
causes FileMetadata to synchronize asset labels and so on.

This adds the IsReadOnly property to FileMetadata to determine
whether it references a file that should not be modified. This
property is used to prevent modification of asset metadata by
the class. In addition, this is used to simplify FileMetadataSet
filtering.

Bug: 150902136
Change-Id: I570456e463f694adada53f039f01f3a5265568df