Fixed unneccesary label changes and duplication.
I241cfe4a08faf2fd919a74c1c19f8bbc03e8ab5e introduce a new label
prefix which preserves labels that would have been stripped out
by old versions of the Version Handler. This change had the
side effect of causing duplicate labels to be written in some
cases.

For example, an asset is labeled with gvhp_target-Editor (valid)
but the label is rewritten as gvh_target-Editor (legacy compatible
format) which causes two labels to be applied to the asset and
an unnecessary update of the asset database.

This change searches the current set of labels applied to an
asset and applies the appropriate prefix based upon the currently
applied labels to minimize the set of labels that are modified.

Bug: 150886091
Change-Id: I43c98064135c8242f283725a62c23257a645c895