Fixed default values being overridden by system settings.
When the ProjectSettings class was configured to read data from
project settings, it was overriding the default value of a setting
with the system setting if the project setting was not present.

This changes the behavior of ProjectSettings to only read and
query keys from the either project settings or system settings by
default. It's still possible to use the fallback behavior if
required.

Change-Id: I5de13a210394993dafe2d247deaabbac34232450