Fixed RunOnMainThread behavior in batch mode vs. executeMethod mode.
When the Unity editor is launched in batch mode,
EditorApplication.update is called periodically just like in
interactive mode. The EditorApplication.update is not called if
the editor is launched with the -executeMethod argument.

This changes RunOnMainThread to only execute jobs in serial if
the editor is launched with the -executeMethod flag. This makes
it possible to get notifications of an asset database refresh
in batch mode without -executeMethod set.

Bug: 150471207
Change-Id: I5878e1f5ff6dc3ee6a8bafd998c54b100ab9f2c5