Fixed some Version Handler issues in 2019.3+
If VersionHandler.UpdateNow() is called from a static
constructor in Unity 2019.3 (which is TBH undefined behavior)
Unity does not save asset metadata changes to disk leaving
all newly _enabled_ plugins in a disabled state.

This changes all entry points to UpdateVersionedAssets() to
run the core logic on the editor thread if possible so that
metadata changes are correctly applied when the asset database
is refreshed.

Other minor clean ups:
- Moved log level loading into a shared method that is used
  by all public entry points.
- Changed calls to UpdateVersionedAssets() from the
  AssetPostProcessor method to excecute after 2s so that
  batch asset changes don't result in multiple calls to
  UpdateVersionedAssets().
- Improved logging of loaded manifests.
- Minor speed up of loaded assembly polling.

Change-Id: I0effaa84c21a13ac5c615520a1a2eb1751d071c5